refactor(packages): move sources into Game package
