+++ src/main/scala/scatan/model/game/Award.scala
@@ -1,4 +1,4 @@
-package scatan.model
+package scatan.model.game
 
 enum AwardType:
   case LongestRoad
