+++ src/test/scala/scatan/model/development/DevelopmentCardTest.scala
@@ -2,7 +2,7 @@ package scatan.model.development
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should
-import scatan.model.{DevelopmentCard, DevelopmentType}
+import scatan.model.game.{DevelopmentCard, DevelopmentType}
 import scatan.BaseTest
 
 class DevelopmentCardTest extends BaseTest:
